reference/netology_pd_diplom/requirements.txt
django
djangorestframework
django-rest-passwordreset
psycopg2-binary
pyyaml
ujson
requests
python-dotenv

pytest==7.2.2
pytest-cov==4.0.0
pytest-django==4.5.2
model-bakery==1.10.3

cachetools==5.3.0
celery==5.2.7
celery[redis]
redis==4.4.0
pytz

gunicorn==20.1.0
